See issue #31: Re-factoring files into separate directories
