+++ package.json
@@ -235,7 +235,7 @@
       },
       {
         "fileMatch": "science.toml",
-        "url": "./schemas/science.schema.json"
+        "url": "https://json.schemastore.org/cargo.json"
       }
     ]
   },
