+++ src/test-explorer/index.ts
@@ -1,0 +1,1 @@
+export * from "./tests-provider";
