+++ src/treeviews/index.ts
@@ -1,0 +1,2 @@
+export * from "./targets-provider";
+export * from "./source-roots-provider";
