+++ src/treeviews/tree-item.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import * as path from "path";
-import { Address } from "./pants/address";
-import { Pants } from "./pants/runner";
+import { Address } from "../pants/address";
+import { Pants } from "../pants/runner";
 
 // The TargetProvider can have several types of tree items.
 // The leaf nodes are the TargetTreeItems, which are the actual targets.
